Refactor semantics package
The previous implementation of AM2 was, in fact, meant to be for
AM3 from Nielsons' book.

This commit fixes that, and also adds operation semantics for the
new instructions added by AM1 and AM3.

Also, fixes a bug with the stack in both machines, where logical
operations would not consume their operands from the stack,
allowing it to grow indefinitely.